Moved all non user stuff out of the users app
This slimming down will make it less likely to be a source of
cyclic dependency issues.

This closes #81
