+++ common/views.py
@@ -4,13 +4,14 @@ from rest_framework.response import Response
 from rest_framework.reverse import reverse
 
 from .models import (
-    Contact, County, Ward, Constituency, ContactType)
+    Contact, County, Ward, Constituency, ContactType, UserCounties)
 
 from .serializers import (
     ContactSerializer, CountySerializer, WardSerializer,
-    ConstituencySerializer, ContactTypeSerializer)
+    ConstituencySerializer, ContactTypeSerializer, InchargeCountiesSerializer)
 
-from .filters import ContactFilter, ConstituencyFilter, WardFilter
+from .filters import (
+    ContactFilter, ConstituencyFilter, WardFilter, UserCountiesFilter)
 
 
 class ContactView(generics.ListCreateAPIView):
@@ -69,6 +70,18 @@ class ContactTypeListView(generics.ListCreateAPIView):
 class ContactTypeDetailView(generics.RetrieveUpdateDestroyAPIView):
     queryset = ContactType.objects.all()
     serializer_class = ContactTypeSerializer
+
+
+class UserCountiesView(generics.ListCreateAPIView):
+    queryset = UserCounties.objects.all()
+    serializer_class = InchargeCountiesSerializer
+    filter_class = UserCountiesFilter
+    ordering_fields = ('user', )
+
+
+class UserCountyDetailView(generics.RetrieveUpdateDestroyAPIView):
+    queryset = UserCounties.objects.all()
+    serializer_class = InchargeCountiesSerializer
 
 
 @api_view(('GET',))
